componentize AppTopBar
