+++ frontend/src/components/layout/AppTopBar/index.tsx
@@ -1,0 +1,3 @@
+export { AppTopBar } from "./AppTopBar";
+export { LogoLink } from "./LogoLink";
+export { ThemeToggleButton } from "./ThemeToggleButton";
